full sync on login
